requirements.txt
langchain
langgraph
streamlit
langchain-google-genai
google-generativeai
requests
